third_party/WebKit/LayoutTests/imported/wpt/service-workers/service-worker/fetch-request-fallback.https-expected.txt

CONSOLE ERROR: XMLHttpRequest cannot load https://127.0.0.1:8444/service-workers/service-worker/resources/fetch-access-control.py?. Response to preflight request doesn't pass access control check: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'https://web-platform.test:8444' is therefore not allowed access.
This is a testharness.js-based test.
FAIL Verify the fallback behavior of FetchEvent assert_unreached: unexpected rejection: https://127.0.0.1:8444/service-workers/service-worker/resources/fetch-access-control.py? should succeed. Reached unreachable code
Harness: the test ran to completion.
